+++ src/nanomodule/datastructures/c_sensordatautm.cpp
@@ -41,17 +41,17 @@ void init_c_sensordatautm(nb::module_& m)
             nb::arg("sensordatalatlon"),
             nb::arg("setutm_zone") = -1)
         .def(
-            nb::init<double, double, int, bool, double, double, double, double, double>(),
+            nb::init<double, double, int, bool, float, float, float, float, float>(),
             DOC(themachinethatgoesping, navigation, datastructures, SensordataUTM, SensordataUTM_5),
             nb::arg("northing")                = 0,
             nb::arg("easting")                 = 0,
             nb::arg("utm_zone")                = 0,
             nb::arg("northern_hemisphere") = true,
-            nb::arg("depth")                   = 0,
-            nb::arg("heave")                   = 0,
-            nb::arg("heading")                 = 0,
-            nb::arg("pitch")                   = 0,
-            nb::arg("roll")                    = 0)
+            nb::arg("depth")                   = 0.0f,
+            nb::arg("heave")                   = 0.0f,
+            nb::arg("heading")                 = 0.0f,
+            nb::arg("pitch")                   = 0.0f,
+            nb::arg("roll")                    = 0.0f)
         .def("__eq__",
              &SensordataUTM::operator==,
              DOC(themachinethatgoesping, navigation, datastructures, SensordataUTM, operator_eq),
